+++ src/main/java/org/jolokia/docker/maven/BuildMojo.java
@@ -11,8 +11,6 @@ import org.jolokia.docker.maven.access.DockerAccessException;
 import org.jolokia.docker.maven.assembly.DockerAssemblyManager;
 import org.jolokia.docker.maven.config.BuildImageConfiguration;
 import org.jolokia.docker.maven.config.ImageConfiguration;
-import org.jolokia.docker.maven.config.LogConfiguration;
-import org.jolokia.docker.maven.config.RunImageConfiguration;
 import org.jolokia.docker.maven.util.MojoParameters;
 
 /**
@@ -30,31 +28,20 @@ public class BuildMojo extends AbstractDockerMojo {
     // Parameters required from Maven when building an assembly. They cannot be injected directly
     // into DockerAssemblyCreator.
     // See also here: http://maven.40175.n5.nabble.com/Mojo-Java-1-5-Component-MavenProject-returns-null-vs-JavaDoc-parameter-expression-quot-project-quot-s-td5733805.html
-    /**
-     * @parameter
-     */
+    /** @parameter */
     private MavenArchiveConfiguration archive;
 
-    /**
-     * @component
-     */
+    /** @component */
     private MavenSession session;
 
-    /**
-     * @component
-     */
+    /** @component */
     private MavenFileFilter mavenFileFilter;
 
-    /**
-     * @component
-     */
+    /** @component */
     private DockerAssemblyManager dockerAssemblyManager;
-    
-    /** @parameter property = "docker.showLogs" default-value="false" */
-    private boolean showLogs;
 
     /**
-     * @parameter default-value="target/docker" property="docker.source.dir"
+     * @parameter default-value="src/main/docker" property="docker.source.dir"
      */
     private String sourceDirectory;
     
@@ -83,17 +70,5 @@ public class BuildMojo extends AbstractDockerMojo {
         dockerAccess.buildImage(imageName, dockerArchive);
     }
 
-    protected boolean showLog(ImageConfiguration imageConfig) {
-        if (showLogs) {
-            return true;
-        }
-        RunImageConfiguration runConfig = imageConfig.getRunConfiguration();
-        if (runConfig != null) {
-            LogConfiguration logConfig = runConfig.getLog();
-            if (logConfig != null) {
-                return logConfig.isEnabled();
-            }
-        }
-        return false;
-    }
+
 }